refactor: remove order package under eatinorders.tobe.domain
