+++ src/main/java/kitchenpos/eatinorders/tobe/domain/EatInOrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain.order;
+package kitchenpos.eatinorders.tobe.domain;
 
 public enum EatInOrderStatus {
     WAITING, ACCEPTED, SERVED, COMPLETED
